Fix: Optimize and Restructure Mentus Repository for Improved Maintainability
Closes #59
