+++ docs/contributing.md
@@ -1,0 +1,51 @@
+# Contributing to Mentus
+
+## Getting Started
+
+Thank you for considering contributing to Mentus! To get started, please follow these steps:
+
+1. Fork the repository on GitHub.
+
+2. Clone your forked repository to your local machine:
+   \`\`\`
+   git clone https://github.com/your-username/mentus.git
+   cd mentus
+   \`\`\`
+
+3. Create a new branch for your feature or bug fix:
+   \`\`\`
+   git checkout -b my-feature-branch
+   \`\`\`
+
+## Setting Up the Development Environment
+
+1. Install dependencies for the backend:
+   \`\`\`
+   cd src/chat-interface/backend
+   npm install
+   \`\`\`
+
+2. Install dependencies for the frontend:
+   \`\`\`
+   cd ../frontend
+   npm install
+   \`\`\`
+
+## Coding Standards
+
+- Follow the existing code style and conventions.
+- Write clear and concise commit messages.
+- Ensure that your code passes all tests and does not introduce new issues.
+
+## Submitting Pull Requests
+
+1. Push your changes to your forked repository:
+   \`\`\`
+   git push origin my-feature-branch
+   \`\`\`
+
+2. Open a pull request on GitHub, providing a clear description of your changes and the problem they solve.
+
+3. Ensure that your pull request passes all checks and is reviewed by a project maintainer.
+
+Thank you for your contributions!
