+++ docs/usage.md
@@ -1,0 +1,33 @@
+# Usage Instructions
+
+## Using the Mentus Extension
+
+1. Open the Mentus extension by clicking on the extension icon in the Chrome toolbar.
+
+2. In the popup window, you can add subjects to the knowledge graph by entering a subject name and clicking the "Add Subject" button.
+
+3. Configure the GraphDB settings by filling in the GraphDB URL, username, and password, and clicking the "Save Configuration" button.
+
+4. Configure the cloud storage settings by filling in the cloud provider, API key, and bucket name, and clicking the "Save Configuration" button.
+
+5. Enter a query in the input field and click the "Send Query" button to get responses from GPT-4o and Claude3Opus.
+
+6. The responses will be displayed in the popup window.
+
+## Viewing the Knowledge Graph
+
+1. The knowledge graph will be displayed in the popup window, showing the subjects and their connections.
+
+2. You can update a subject's connections by entering the subject name and the new connections, and clicking the "Update Subject" button.
+
+3. You can delete a subject by entering the subject name and clicking the "Delete Subject" button.
+
+## Troubleshooting
+
+If you encounter any issues while using the Mentus extension, please refer to the following troubleshooting steps:
+
+1. Ensure that you have entered the correct GraphDB and cloud storage configurations.
+
+2. Check the browser console for any error messages.
+
+3. If the issue persists, please contact the Mentus support team for assistance.
