+++ src/chat-interface/frontend/src/App.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import './App.css';
+import './styles/App.css';
 
 function App() {
   const [messages, setMessages] = useState([]);
