+++ src/chat-interface/webpack.config.js
