+++ src/chrome-extension/manifest.json
