+++ src/chrome-extension/package.json
