+++ src/chrome-extension/src/background/background.js
