+++ src/chrome-extension/src/popup/popup.html
@@ -48,6 +48,6 @@
   <button type="submit">Save Configuration</button>
 </form>
 <div id="response"></div>
-  <script src="popup.js"></script>
+  <script src="src/popup/popup.js"></script>
 </body>
 </html>
