+++ tests/chrome-extension/test_cloud_storage.html
