+++ tests/chrome-extension/test_knowledge_graph.js
